Add test_util.py
test.py is now 1500 lines long and is turning into dumping ground.  Add
a separate module for util unit tests.
